refactor: move publish controls to `reaction-catalog` package
