+++ imports/plugins/included/product-detail-simple/client/containers/publish.js
@@ -4,7 +4,7 @@ import { registerComponent, composeWithTracker } from "@reactioncommerce/reactio
 import { Meteor } from "meteor/meteor";
 import { ReactionProduct } from "/lib/api";
 import { Products } from "/lib/collections";
-import PublishContainer from "/imports/plugins/core/revisions/client/containers/publishContainer";
+import PublishContainer from "/imports/plugins/core/catalog/client/containers/publishContainer";
 
 class ProductPublishContainer extends Component {
   handleMetaRemove = (productId, metafield) => {
